Remove fast output, as it was poorly designed.
The specific dprintf is never directly called by client code.  Since the
client may need to swap to the slow dprintf, depending on what is being
logged, the design doesn't allow for easy swapping, except to reset the
logger.  Setting the logger isn't thread safe, and the swapping is not
ideal, so it is best to just remove the fast logger and always use the
utf8-correct  logger.